.gitlab-ci.yml: ensure that tagged pipelines only run with fixed docker tags
We had a chicken-and-egg problem in this project. Therefore the RELEASE_DOCKER_TAG variable is introduced.

If a tagged pipeline is running we need to ensure that every component has a fixed version.
This ensures that we can rebuild a release in the future.
Also the default branch builds with this tag to test the build before a tag is created.

Only the test builds are running with the 'latest' tag.

Signed-off-by: Christoph Krutz <christoph.krutz@tq-group.com>