Makefile: Fix build with docker compose
'docker-compose' (v1) and 'docker compose' (v2) handles the image dependencies differently.
'docker-compose' (v1) resolves the dependencies, 'docker compose' (v2) doesn't.
A target for every image is added to define the image dependencies.

Remove the predefined CI variables as the local builds shall be independent from the CI jobs.
The ALL_IMAGES variable is not used anymore, the DOCKER_BASE_TAG variable is replaced with
the BUILD_TAG variable.

The .NOTPARALLEL flag is removed as it doesn't work properly, the flag causes all targets to be
processed serially. The recursive use of make handles the serialization of the targets, now.

Also existing certificates can be used, now.

Signed-off-by: Christoph Krutz <christoph.krutz@tq-group.com>